requirements.txt
aiofiles==0.6.0
aioredis==1.3.1
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
black==20.8b1
blinker==1.4
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
distro==1.5.0
docker==4.4.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
h11==0.12.0
h2==4.0.0
hiredis==1.1.0
hpack==4.0.0
Hypercorn==0.11.1
hyperframe==6.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mypy-extensions==0.4.3
paramiko==2.7.2
pathspec==0.8.1
pkg-resources==0.0.0
priority==1.3.0
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.15.0
PyYAML==5.3.1
Quart==0.14.1
Quart-Session==1.0.2
regex==2020.11.13
requests==2.25.1
six==1.15.0
texttable==1.6.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
wsproto==1.0.0
